+++ packages/root/src/methods.ts
@@ -1,3 +1,0 @@
-export const PUSH = "push";
-export const REPLACE = "replace";
-export const ANCHOR = "anchor";