+++ packages/root/types/index.d.ts
@@ -1,4 +1,4 @@
-export * from './types';
-import Common from './common';
-import { PUSH, REPLACE, ANCHOR } from "./methods";
-export { Common, PUSH, REPLACE, ANCHOR };
+export * from "./types";
+import Common from "./common";
+import { PUSH, REPLACE, ANCHOR, POP } from "./constants";
+export { Common, PUSH, REPLACE, ANCHOR, POP };