+++ packages/root/types/methods.d.ts
@@ -1,3 +1,0 @@
-export declare const PUSH = "push";
-export declare const REPLACE = "replace";
-export declare const ANCHOR = "anchor";