+++ packages/root/types/types/index.d.ts
@@ -1,5 +1,5 @@
 export { LocationDetails, PartialLocation, HickoryLocation, AnyLocation } from "./location";
-export { History, ToArgument, Action, NavType, PendingNavigation, ResponseHandler, Options, CommonHistory } from "./hickory";
+export { History, ToArgument, Action, NavType, PendingNavigation, ResponseHandler, Options, CommonHistory, PushType, ReplaceType, PopType, AnchorType } from "./hickory";
 export { KeyMethods, KeyFns } from "./keygen";
 export { QueryFunctions, LocationFactoryOptions, LocationMethods } from "./locationFactory";
 export { NavigationInfo, ConfirmationFunction, ConfirmationMethods } from "./navigationConfirmation";