+++ packages/root/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
-* Export `PUSH`, `REPLACE`, and `ANCHOR` constants, move `Common` to a named export.
+* Export `PUSH`, `REPLACE`, `ANCHOR`, and `POP` constants.
+* Move `Common` to a named export.
 
 ## 1.0.2